updates for final format.
the suggested nameing scheme was not followed
so this commit addes support for parsing the final document
and also updates the data.csv test file with a copy of the
final data.

to make this more useful the tag system has been adjusted to
generate tags from titles delimited by white space.

this change also disables geenrating indigivual session per room or tag
and instead only  genergets the aggregate icals but alls adds generation of
indivigual icals for all sessions.